omachemainproj/requirements.txt
dj-rest-auth
Django
django-allauth
django-oauth-toolkit
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-unfold
